refactor: extract recording form component
